+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY4TS7NVFE2H4A4NZK9VVM8Z"
+        "01KY524HNDQHMKDBXSN3A5X2XV"
     )
     r"""The unique identifier of the data source"""