+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KY4TS7GJK3ETPVFFDAA563D8"
+    id: Optional[str] = "01KY524HE6GNJ7NDJ7DXST39ZY"
 
     description: Optional[str] = None
 
@@ -1362,7 +1362,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY4TS7G50MQ38E3XCBMRT15E"
+        "tool_01KY524HDMJ1JW0B710XV4E7QR"
     )
 
     display_name: Optional[str] = None
@@ -1493,7 +1493,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KY4TS7G4D2MQWA69HVXJE9H5"
+    id: Optional[str] = "01KY524HDDWXR7PW8F851635ZV"
 
     description: Optional[str] = None
 
@@ -1637,7 +1637,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY4TS7G2FPQC0SAMW7VGZ83K"
+        "tool_01KY524HD8GYVR5EXVRS5G0PX4"
     )
 
     display_name: Optional[str] = None
@@ -1942,7 +1942,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY4TS7FZ6N43QFPKKYN24G4X"
+        "tool_01KY524HD3CDKN6W9KPY15Y2C9"
     )
 
     display_name: Optional[str] = None
@@ -2136,7 +2136,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY4TS7FX9TEC7C85QB5JEHS5"
+        "tool_01KY524HCX2XZKX0YFWR490ZMF"
     )
 
     display_name: Optional[str] = None
@@ -2334,7 +2334,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KY4TS7FTW8B8DKPJMCKCNZJE"
+        "tool_01KY524HCSVNJ2VQ9V71JA681D"
     )
 
     display_name: Optional[str] = None