+++ packages/orq-rc/RELEASES.md
@@ -9649,3 +9649,13 @@ Based on:
 - [python v4.12.0-rc.53] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.53 - packages/orq-rc
+
+## 2026-07-22 14:05:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.54] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.54 - packages/orq-rc